requirements.txt
Django>=5.2.0
djangorestframework>=3.16.0
pandas>=2.2.0
scikit-learn>=1.6.0
nltk>=3.9.0
matplotlib>=3.10.0
numpy>=2.2.0
joblib>=1.4.0
xhtml2pdf>=0.2.17
reportlab>=4.4.0
python-dateutil>=2.9.0
pytz>=2025.2
requests>=2.32.0
pillow>=11.2.0
channels>=4.0.0
channels-redis>=4.1.0
whitenoise>=6.6.0
redis>=5.0.0
psycopg2-binary>=2.9.9
django-debug-toolbar>=4.3.0
celery>=5.3.0

DB_NAME=your_database_name
DB_USER=your_database_user
DB_PASSWORD=your_database_password
DB_HOST=localhost
DB_PORT=5432
DJANGO_SECRET_KEY=your_secret_key
DJANGO_DEBUG=True
DJANGO_ALLOWED_HOSTS=localhost,127.0.0.1
REDIS_HOST=localhost
REDIS_PORT=6379 